Split AbstractApp from ConsoleApp
