+++ src/app/consoleApp.js
@@ -1,49 +1,27 @@
 "use strict";
 
-var spawn        = require("child_process").spawn;
-var EventEmitter = require('events').EventEmitter;
-var LineEventEmitter = require("../utils/lineEventEmitter");
+var AbstractApp      = require("./abstractApp");
 var CommandResult    = require("../commandResult");
+var EventEmitter     = require('events').EventEmitter;
 
 function ConsoleApp(cmd, args, cwd) {
   this.cmd = cmd;
   this.args = this.normalizeArgs(args);
   this.cwd = cwd;
 
-  this.emitter = new EventEmitter();
-
   this._input = [];
   this._expected = [];
-  this.exitCode = null;
 
-  this._consoleOut = false;
+  this.emitter = new EventEmitter();
 }
 
-ConsoleApp.prototype.normalizeArgs = function(args) {
-  var ret = [];
-  if (args) {
-    for (var i=0; i<args.length; i++) {
-      if (Array.isArray(args[i])) {
-        ret.concat(args[i]);
-      } else {
-        ret.push(args[i]);
-      }
-    }
-  }
-  return ret;
-};
-
-ConsoleApp.prototype.getCommandLine = function() {
-  return this.cmd + " " + this.args.join(" ");
-};
+ConsoleApp.prototype = new AbstractApp();
 
 ConsoleApp.prototype.input = function() {
   if (arguments.length === 0) {
     return [].concat(this._input);
   }
-  for (var i=0; i<arguments.length; i++) {
-    this._input.push(arguments[i]);
-  }
+  this._input = this._input.concat(this.normalizeArgs(arguments));
   return this;
 };
 
@@ -51,78 +29,35 @@ ConsoleApp.prototype.expected = function() {
   if (arguments.length === 0) {
     return [].concat(this._expected);
   }
-  for (var i=0; i<arguments.length; i++) {
-    this._expected.push(arguments[i]);
-  }
+  this._expected = this._expected.concat(this.normalizeArgs(arguments));
   return this;
 };
 
-ConsoleApp.prototype.consoleOut = function() {
-  if (arguments.length === 0) {
-    return this._consoleOut;
-  } else {
-    this._consoleOut = arguments[0];
-    return this;
+ConsoleApp.prototype.doClose = function(code) {
+  if (this._consoleOut) {
+    process.stdout.write("codecheck: Finish '" + this.getCommandLine() + "with code " + code + "\n");
   }
 };
 
-ConsoleApp.prototype.getExitCode = function() {
-  return this.exitCode;
-};
-
-ConsoleApp.prototype.run = function(additionalArgs) {
-  var self = this;
-  var args = this.args.concat(this.normalizeArgs(additionalArgs));
-  var options = {
-    env: process.env
-  };
-  if (this.cwd) {
-    options.cwd = this.cwd;
-  }
-
-  var emitter = this.emitter;
-  var stdoutBuf = new LineEventEmitter(emitter, "stdout");
-  var stderrBuf = new LineEventEmitter(emitter, "stderr");
-
-  var p = spawn(this.cmd, args, options);
-  p.stdout.on("data", function(data) {
-    if (self._consoleOut) {
-      process.stdout.write(data);
-    }
-    stdoutBuf.add(data);
-  });
-  p.stderr.on("data", function(data) {
-    if (self._consoleOut) {
-      process.stderr.write(data);
-    }
-    stderrBuf.add(data);
-  });
-  p.on('close', function(code) {
-    stdoutBuf.close();
-    stderrBuf.close();
-
-    self.executed = true;
-    self.exitCode = code;
-    self.childProcess = null;
-
-    emitter.emit("end", code);
-    if (self._consoleOut) {
-      process.stdout.write("codecheck: Finish '" + self.getCommandLine() + "with code " + code + "\n");
-    }
-  });
-  p.stdin.setEncoding("utf-8");
-  this.childProcess = p;
-  var values = self.input();
+ConsoleApp.prototype.doRun = function(process) {
+console.log("doRun");
+  process.stdin.setEncoding("utf-8");
+  var values = this.input();
   while (values.length) {
     var value = values.shift();
-    p.stdin.write(value + "\n");
+    process.stdin.write(value + "\n");
   }
-  p.stdin.end();
+  process.stdin.end();
 };
 
-ConsoleApp.prototype.runAndVerify = function(done) {
+ConsoleApp.prototype.runAndVerify = function(additionalArgs, done) {
+  if (typeof(additionalArgs) === "function") {
+    done = additionalArgs;
+    additionalArgs = null;
+  }
   var errors = [];
   var values = this.expected();
+console.log("verify", values);
   this.onStdout(function(data) {
     if (values.length === 0) {
       errors.push("Expected vlaue is nothing, but output is " + data);
@@ -146,30 +81,11 @@ ConsoleApp.prototype.runAndVerify = function(done) {
       result = result.withErrors(errors);
     }
     if (done) {
+console.log("Done", result);
       done(result);
     }
   });
-  this.run();
-};
-
-//Events
-ConsoleApp.prototype.on = function(name, callback) {
-  if (!name || !callback) {
-    throw new Error("Illegal arguments: name=" + name + ", callback=" + callback);
-  }
-  this.emitter.on(name, callback);
-};
-
-ConsoleApp.prototype.onEnd = function(callback) {
-  this.on("end", callback);
-};
-
-ConsoleApp.prototype.onStdout = function(callback) {
-  this.on("stdout", callback);
-};
-
-ConsoleApp.prototype.onStderr = function(callback) {
-  this.on("stderr", callback);
+  this.run(additionalArgs);
 };
 
 module.exports = ConsoleApp;
